+++ modules/tinymce/src/core/main/ts/api/Editor.ts
@@ -239,6 +239,7 @@ class Editor implements EditorObservable {
   public toggleNativeEvent: EditorObservable['toggleNativeEvent'];
   public unbindAllNativeEvents: EditorObservable['unbindAllNativeEvents'];
   public fire: EditorObservable['fire'];
+  public dispatch: EditorObservable['dispatch'];
   public on: EditorObservable['on'];
   public off: EditorObservable['off'];
   public once: EditorObservable['once'];
@@ -323,7 +324,7 @@ class Editor implements EditorObservable {
     this.mode = createMode(self);
 
     // Call setup
-    editorManager.fire('SetupEditor', { editor: this });
+    editorManager.dispatch('SetupEditor', { editor: this });
     const setupCallback = Options.getSetupCallback(self);
     if (Type.isFunction(setupCallback)) {
       setupCallback.call(self, self);
@@ -613,7 +614,7 @@ class Editor implements EditorObservable {
       }
 
       self.load();
-      self.fire('show');
+      self.dispatch('show');
     }
   }
 
@@ -642,7 +643,7 @@ class Editor implements EditorObservable {
       }
 
       self.hidden = true;
-      self.fire('hide');
+      self.dispatch('hide');
     }
   }
 
@@ -675,7 +676,7 @@ class Editor implements EditorObservable {
    * tinymce.activeEditor.setProgressState(true, 3000);
    */
   public setProgressState(state: boolean, time?: number) {
-    this.fire('ProgressState', { state, time });
+    this.dispatch('ProgressState', { state, time });
   }
 
   /**
@@ -705,7 +706,7 @@ class Editor implements EditorObservable {
       args.element = elm;
 
       if (!args.no_events) {
-        self.fire('LoadContent', args);
+        self.dispatch('LoadContent', args);
       }
 
       args.element = elm = null;
@@ -738,12 +739,12 @@ class Editor implements EditorObservable {
     html = args.content = self.getContent(args);
 
     if (!args.no_events) {
-      self.fire('SaveContent', args);
+      self.dispatch('SaveContent', args);
     }
 
     // Always run this internal event
     if (args.format === 'raw') {
-      self.fire('RawSaveContent', args);
+      self.dispatch('RawSaveContent', args);
     }
 
     html = args.content;
@@ -778,6 +779,8 @@ class Editor implements EditorObservable {
   /**
    * Sets the specified content to the editor instance, this will cleanup the content before it gets set using
    * the different cleanup rules options.
+   * <br>
+   * <em>Note: The content return value was deprecated in TinyMCE 6.0 and has been marked for removal in TinyMCE 7.0.</em>
    *
    * @method setContent
    * @param {String} content Content to set to editor, normally HTML contents but can be other formats as well.
@@ -900,7 +903,7 @@ class Editor implements EditorObservable {
     this.isNotDirty = !state;
 
     if (state && state !== oldState) {
-      this.fire('dirty');
+      this.dispatch('dirty');
     }
   }