+++ modules/tinymce/src/core/main/ts/selection/EventProcessRanges.ts
@@ -8,7 +8,7 @@
 import { Arr } from '@ephox/katamari';
 
 const processRanges = (editor, ranges: Range[]): Range[] => Arr.map(ranges, (range) => {
-  const evt = editor.fire('GetSelectionRange', { range });
+  const evt = editor.dispatch('GetSelectionRange', { range });
   return evt.range !== range ? evt.range : range;
 });