+++ modules/tinymce/src/plugins/autoresize/main/ts/api/Events.ts
@@ -9,7 +9,7 @@ import Editor from 'tinymce/core/api/Editor';
 import { EditorEvent } from 'tinymce/core/api/util/EventDispatcher';
 
 const fireResizeEditor = (editor: Editor): EditorEvent<{}> =>
-  editor.fire('ResizeEditor');
+  editor.dispatch('ResizeEditor');
 
 export {
   fireResizeEditor